refactor: move tests to tests.rs
